+++ Scr/Components/Player.h
@@ -16,20 +16,20 @@ namespace booble
 	class Player : public boop::Component, public boop::Observer, public boop::Subject
 	{
 	public:
-		Player(boop::GameObject* owner, float speed);
+		Player(boop::GameObject* owner, int speed);
 		~Player() = default;
 
 		void FixedUpdate() override;
 		void Update() override;
 		void Render() const override;
 
-		void ChangeSpeed(float speed) { m_Speed = speed; }
-		float GetSpeed() const { return m_Speed; }
+		void ChangeSpeed(int speed) { m_Speed = speed; }
+		int GetSpeed() const { return m_Speed; }
 
 		int GetScore() const { return m_Score; }
 		void OnNotify(boop::Event event) override;
 	private:
-		float m_Speed{};
+		int m_Speed{};
 		int m_Score{};
 	};
